Упрощает логику PageController и добавляет MovieController
