+++ src/controllers/page.js
@@ -1,5 +1,4 @@
-import PopupComponent from '../components/popup.js';
-import FilmCardComponent from '../components/film-card.js';
+// import PopupComponent from '../components/popup.js';
 import ShowMoreButtonComponent from '../components/show-more-button.js';
 import FilmsListComponent from '../components/films-list.js';
 import NoFilmsComponent from '../components/no-films.js';
@@ -7,9 +6,10 @@ import SpecialListComponent from '../components/special-list.js';
 import SortComponent from '../components/sort.js';
 import SiteMenuComponent from '../components/menu.js';
 import RatingComponent from '../components/rating.js';
-import {isEscEvent, siteMainElement, sortByRating, sortByComments, getRating} from '../utils/common.js';
+import {siteMainElement, sortByRating, sortByComments, getRating} from '../utils/common.js';
 import {render, RenderPosition, remove} from '../utils/render.js';
 import {SHOWING_CARDS_COUNT_ON_START, SHOWING_CARDS_COUNT_BY_BUTTON, CARDS_COUNT_ADDITIONAL} from '../const.js';
+import MovieController from './movie.js';
 
 export default class PageController {
   constructor(container) {
@@ -57,12 +57,6 @@ export default class PageController {
     render(siteMainElement, new SiteMenuComponent(filtersCounts), RenderPosition.BEFOREEND);
   }
 
-  _renderCard(card, container) {
-    const newCard = new FilmCardComponent(card);
-    render(container, newCard, RenderPosition.BEFOREEND);
-    this._generateDetailedCard(card, newCard);
-  }
-
   /**
    * Рендерит карточки фильмов
    * @param  {number} count     кол-во карточек
@@ -72,7 +66,8 @@ export default class PageController {
   _renderCards(count, container, array) {
     const currentCardsList = array.slice(0, count);
     currentCardsList.forEach((card) => {
-      this._renderCard(card, container);
+      const movieController = new MovieController(container);
+      movieController.render(card);
     });
   }
 
@@ -80,45 +75,45 @@ export default class PageController {
    * Отрисовывает попап
    * @param  {object} popup объект попапа
    */
-  _renderPopup(popup) {
-    const body = document.querySelector(`body`);
+  // _renderPopup(popup) {
+  //   const body = document.querySelector(`body`);
 
-    const onEscPress = (evt) => {
-      if (isEscEvent(evt)) {
-        remove(popup);
-        document.removeEventListener(`keydown`, onEscPress);
-      }
-    };
+  //   const onEscPress = (evt) => {
+  //     if (isEscEvent(evt)) {
+  //       remove(popup);
+  //       document.removeEventListener(`keydown`, onEscPress);
+  //     }
+  //   };
 
-    const onCloseButtonClick = () => {
-      remove(popup);
-      document.removeEventListener(`keydown`, onEscPress);
-    };
+  //   const onCloseButtonClick = () => {
+  //     remove(popup);
+  //     document.removeEventListener(`keydown`, onEscPress);
+  //   };
 
-    if (body.querySelector(`.film-details`)) {
-      body.querySelector(`.film-details`).remove();
-    }
+  //   if (body.querySelector(`.film-details`)) {
+  //     body.querySelector(`.film-details`).remove();
+  //   }
 
-    render(body, popup, RenderPosition.BEFOREEND);
-    document.addEventListener(`keydown`, onEscPress);
+  //   render(body, popup, RenderPosition.BEFOREEND);
+  //   document.addEventListener(`keydown`, onEscPress);
 
-    popup.bind(onCloseButtonClick);
-  }
+  //   popup.bind(onCloseButtonClick);
+  // }
 
   /**
    * Генерирует попап при клике на элементы карточки
    * @param  {object} card    объект карточки фильма
    * @param  {object} newCard компонент карточки фильма
    */
-  _generateDetailedCard(card, newCard) {
-    const detailedCard = new PopupComponent(card);
+  // _generateDetailedCard(card, newCard) {
+  //   const detailedCard = new PopupComponent(card);
 
-    const onCardClick = () => {
-      this._renderPopup(detailedCard);
-    };
+  //   const onCardClick = () => {
+  //     this._renderPopup(detailedCard);
+  //   };
 
-    newCard.bind(onCardClick);
-  }
+  //   newCard.bind(onCardClick);
+  // }
 
   /**
    * Рендерит кнопку Show more
@@ -156,7 +151,8 @@ export default class PageController {
     const container = parentContainer.querySelector(`.${title.substring(0, 3).toLowerCase()}-container`);
     const currentCardsList = arr.slice(0, CARDS_COUNT_ADDITIONAL);
     currentCardsList.forEach((card) => {
-      this._renderCard(card, container);
+      const movieController = new MovieController(container);
+      movieController.render(card);
     });
   }
 
@@ -192,7 +188,8 @@ export default class PageController {
 
       cards.slice(prevCardsCount, this._showingCardsCount)
       .forEach((card) => {
-        this._renderCard(card, this._filmsCardsContainer);
+        const movieController = new MovieController(this._filmsCardsContainer);
+        movieController.render(card);
       });
 
       if (this._showingCardsCount >= cards.length) {
